pyproject.toml
[project]
name = "temporal-graph-rag"
version = "0.1.0"
description = "Temporal Graph RAG - фреймворк для построения временных графов знаний"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.0",
    "scikit-learn>=1.3.0",
    "httpx>=0.27.0",
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "transformers>=4.30.0",
    "pyarrow>=15.0.0",
    "faiss-cpu>=1.8.0",
    "pandas>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["temporal_graph_rag"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
